requirements.txt
apscheduler
python-telegram-bot
requests
selenium
webdriver-manager
pytz
